Some refactoring

+++ src/main/java/ru/skypro/homework/controller/CommentsController.java
@@ -11,22 +11,22 @@ import ru.skypro.homework.dto.CreateOrUpdateCommentDto;
 @CrossOrigin(value = "http://localhost:3000")
 public class CommentsController {
     @GetMapping("/{id}/comments")
-    public ResponseEntity<CommentsDto> getComments(@PathVariable Long id) {
+    public ResponseEntity<CommentsDto> getComments(@PathVariable Integer id) {
         return ResponseEntity.ok(null);
     }
 
     @PostMapping("/{id}/comments")
-    public ResponseEntity<CommentDto> addComment(@PathVariable Long id, @RequestBody CreateOrUpdateCommentDto dto) {
+    public ResponseEntity<CommentDto> addComment(@PathVariable Integer id, @RequestBody CreateOrUpdateCommentDto dto) {
         return ResponseEntity.ok(null);
     }
     @DeleteMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<Void> deleteComment(@PathVariable Long adId, @PathVariable Long commentId) {
+    public ResponseEntity<Void> deleteComment(@PathVariable Integer adId, @PathVariable Integer commentId) {
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<CommentDto> updateComment(@PathVariable Long adId,
-                                                    @PathVariable Long commentId,
+    public ResponseEntity<CommentDto> updateComment(@PathVariable Integer adId,
+                                                    @PathVariable Integer commentId,
                                                     @RequestBody CreateOrUpdateCommentDto dto) {
         return ResponseEntity.ok(null);
     }
